add disabledBtn
