+++ src/partials/basket.html
@@ -41,8 +41,8 @@
   <div>
     <ul class="modal-basket__list list">
     <li class="modal-basket__item"><button class="modal-basket__continue" type="button">CONTINUE  BOOKING</button></li>
-    <li class="modal-basket__item"><button class="modal-basket__buy" type="button">BUY</button></li>
-    <li class="modal-basket__item"><button class="modal-basket__clear" type="button">CLEAR THE LIST</button></li>
+    <li class="modal-basket__item"><button class="modal-basket__buy" type="button" disabled>BUY</button></li>
+    <li class="modal-basket__item"><button class="modal-basket__clear" type="button" disabled>CLEAR THE LIST</button></li>
   </ul>
   </div>
 </div>
